Move compaction code to compaction package
